Déplacement des composants de la page QuiSuisJe, Correctif des imports et Refactorisation
