+++ client/src/pages/Public/QuiSuisJePage/QuiSuisJePage.tsx
@@ -1,9 +1,9 @@
-import NavRoot from "../../components/NavRoot/NavRoot";
-import style from "../stylePage.module.css";
-import QuiSuisJeRoot from "../../components/QuiSuisJeRoot/QuiSuisJeRoot";
-import FooterRoot from "../../components/FooterRoot/FooterRoot";
-import Helmet_QuiSuisJe_SEO from "../../seo/QuiSuisJeSEO/Helmet_QuiSuisJe.seo";
-import useScrollToHash from "../../hook/useScrollToHash";
+import style from "../../stylePage.module.css";
+import NavRoot from "../../../components/NavRoot/NavRoot";
+import QuiSuisJeRoot from "../../../components/Public/QuiSuisJeRoot/QuiSuisJeRoot";
+import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import Helmet_QuiSuisJe_SEO from "../../../seo/QuiSuisJeSEO/Helmet_QuiSuisJe.seo";
+import useScrollToHash from "../../../hook/useScrollToHash";
 
 function QuiSuisJePage() {
     useScrollToHash() /* Utilisation du hook pour gérer le scroll vers les sections avec des IDs */
